backend/requirements.txt
# Optional only; the local service runs on the Python standard library.
pypdf==5.4.0
pypdfium2>=4.0
python-docx>=1.0
python-pptx>=1.0
openai
